% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate_fns.R
\name{ate}
\alias{ate}
\alias{naive_ate}
\alias{ipw1_ate}
\alias{ipw2_ate}
\alias{ipw3_ate}
\alias{strat_ate}
\alias{rf_ate}
\alias{cart_ate}
\alias{pscart_ate}
\alias{strat_regr_ate}
\alias{match_ps_ate}
\alias{match_prog_ate}
\alias{match_both_ate}
\alias{caliper_ps_ate}
\alias{caliper_prog_ate}
\alias{caliper_both_ate}
\alias{caliper_nn_ate}
\alias{caliper_logit_ate}
\alias{regr_ate}
\alias{dr_ate}
\alias{bal_ate}
\alias{highdim_bal_ate}
\alias{cem_ate}
\alias{grf_ate}
\title{Functions to estimate the average treatment effect}
\usage{
ate(this_data, specific_ate, ...)

naive_ate(this_data, ...)

ipw1_ate(this_data, ...)

ipw2_ate(this_data, ...)

ipw3_ate(this_data, ...)

strat_ate(this_data, ...)

rf_ate(this_data, outcome_fm, ...)

cart_ate(this_data, outcome_fm, ...)

pscart_ate(this_data, ps_fm, ...)

strat_regr_ate(this_data, outcome_fm, outcome_fam = gaussian, ...)

match_ps_ate(this_data, ...)

match_prog_ate(this_data, ...)

match_both_ate(this_data, ...)

caliper_ps_ate(this_data, ...)

caliper_prog_ate(this_data, ...)

caliper_both_ate(this_data, ...)

caliper_nn_ate(this_data, ...)

caliper_logit_ate(this_data, ...)

regr_ate(this_data, outcome_fm, outcome_fam = gaussian, ...)

dr_ate(this_data, outcome_fm, outcome_fam = gaussian, ...)

bal_ate(this_data, cov_ids, ...)

highdim_bal_ate(this_data, cov_ids, ...)

cem_ate(this_data, cov_ids, ...)

grf_ate(this_data, cov_ids, ...)
}
\arguments{
\item{this_data}{a data frame with value \code{y} for the outcome, \code{d} for the treatment, and \code{w} for a case weight.}

\item{...}{additional arguments to be passed indicating, for example, the outcome formula, propensity score formula, or other things on a case by case basis.}
}
\value{
a 1 x 1 data frame containing the named ATE
}
\description{
Functions to estimate the average treatment effect
}
\examples{

ate_list <- list(
ipw1 = ipw1_ate,
ipw2 = ipw2_ate,
ipw3 = ipw3_ate,
strat = strat_ate,
strat_regr = strat_regr_ate,
match_ps = match_ps_ate,
match_prog = match_prog_ate,
match_both = match_both_ate,
cal_ps = caliper_ps_ate,
cal_both = caliper_both_ate,
regr = regr_ate,
dr = dr_ate,
bal = bal_ate,
hdbal = highdim_bal_ate
)
gen_mod <- generate_data(n = 100, 
                         dgp = 'ks', 
                         correct_outcome = FALSE,
                         correct_ps = TRUE)
this_data <- gen_mod$data
this_data <- estimate_scores(this_data, outcome_fm = outcome_fm,
                             ps_fm = ps_fm,
                             ps_fam = ps_fam,
                             outcome_fam = outcome_fam)
thetahat <- with(gen_mod,
                 estimate_ates(this_data,
                          ate_list,
                          cov_ids = cov_ids,
                          outcome_fm = stringr::str_c('d + ', outcome_fm),
                          outcome_fam = outcome_fam))
thetahat

}
