% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_fns.R
\name{cace}
\alias{cace}
\alias{iv_fn}
\alias{ivw_fn}
\alias{at_fn}
\alias{pp_fn}
\alias{tsls_fn}
\alias{itt_regr_fn}
\alias{atregr_fn}
\alias{ppregr_fn}
\alias{ivs_fn}
\alias{ats_fn}
\alias{pps_fn}
\alias{ipw_fn}
\title{Functions to estimate the complier average causal effect or local average treatment effect}
\usage{
cace(ds, specific_cace, ...)

iv_fn(ds)

ivw_fn(ds)

at_fn(ds)

pp_fn(ds)

tsls_fn(ds)

itt_regr_fn(ds)

atregr_fn(ds)

ppregr_fn(ds)

ivs_fn(ds)

ats_fn(ds)

pps_fn(ds)

ipw_fn(ds)
}
\arguments{
\item{ds}{a data frame with value \code{y} for the outcome, \code{z} for randomization/instrument, \code{s} for treatment receipt, and covariate \code{x}.}

\item{...}{additional arguments to be passed.}
}
\value{
a 1 x 1 data frame containing the named ATE
}
\description{
Functions to estimate the complier average causal effect or local average treatment effect
}
\examples{



}
