% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_combination.R
\name{do_combination}
\alias{do_combination}
\title{Workhorse function for creating a synthetic estimator by combining multiple candidate estimators.}
\usage{
do_combination(
  ests,
  name_0,
  C,
  print = FALSE,
  exclude_t0 = FALSE,
  bias_type = "raw_diff",
  boot_mean = NULL,
  ate_0 = NULL,
  n = NULL
)
}
\arguments{
\item{ests}{one-row, p-column data frame of estimators.}

\item{name_0}{character value of the name of the presumed unbiased estimator, \eqn{\theta_0}. Default is NULL, which returns results for using each candidate estimator as \eqn{\theta_0}, one synthetic estimator for each. If \code{ate_0} is given, then it is used as \eqn{\theta_0} in place of this.}

\item{C}{p x p covariance matrix of \code{ests}.}

\item{print}{logical indicating whether details should be printed. Default is FALSE.}

\item{exclude_t0}{logical indicating whether \eqn{\theta_0} should be considered an external estimator (not a candidate for combining with others). Default is FALSE.}

\item{bias_type}{method to compute the bias in the mean squard error. Default is \code{raw_diff}, which computes the bias as the raw difference between each of the candidate estimator and \eqn{\theta_0}. Other options to compute the bias include: \code{bootstrap} which computes the bias as the difference between the mean of the bootstrap samples and the observed value of \eqn{\theta_0}; \code{bootstrap_all} which computes the bias as the mean of the difference between the bootstrapped version of the candidate estimator and the bootstrapped version of \eqn{\theta_0}; \code{none} which assumes no bias; \code{shrunk} which computes the bias as the raw difference divided by \code{n}.}

\item{boot_mean}{mean of bootstrap samples for bootstrap-based bias estimation.}

\item{ate_0}{external value of \eqn{\theta_0}. Default is NULL, in which case \eqn{\theta_0} is taken to be \code{name_0}.}

\item{n}{sample size. Default is NULL. Needed only if \code{bias_type} is \code{shrunk}.}
}
\value{
list of three objects, including \code{ate_res} which gives results for the synthetic estimator, \code{b_res} which gives results for how the estimators were combined, and \code{C} which gives the covariance matrix of the estimators.
}
\description{
Creates a synthetic estimator by minimizing the (estimated) mean squared error of a linear combination of multiple candidate estimators.
}
