% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ates.R
\name{estimate_ates}
\alias{estimate_ates}
\title{Estimate many average treatment effects}
\usage{
estimate_ates(this_data, fns, profile = FALSE, ...)
}
\arguments{
\item{this_data}{data frame to estimate from.}

\item{fns}{list of functions to compute ATEs.}

\item{profile}{logical indicating whether the computation times should be returned rather than the estimates themselves.}

\item{...}{additional arguments to be passed to the functions in \code{fns}}
}
\value{
a data frame containing all the ATEs, one per column.
}
\description{
Estimate many average treatment effects
}
\examples{


}
