% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_scores.R
\name{estimate_scores}
\alias{estimate_scores}
\title{Estimate propensity and prognostic scores.}
\usage{
estimate_scores(
  this_data,
  outcome_fm,
  outcome_fam = gaussian,
  ps_fm,
  ps_fam = binomial,
  wt = NULL
)
}
\arguments{
\item{this_data}{data frame to estimate scores on.}

\item{outcome_fm}{formula for the outcome model.}

\item{outcome_fam}{distributional family for outcome model.}

\item{ps_fm}{formula for the propensity score model.}

\item{ps_fam}{distributional family for propensity score model.}

\item{wt}{optional case weights. Default is NULL.}
}
\value{
data frame augmented with \code{ps}, the propensity score, \code{prog_score}, the prognostic score, \code{pi}, the propensity score if treated or 1 minus the propensity score if control, and \code{ps_strat}, a numeric value indicating the quintile of the propensity score that each individual falls in.
}
\description{
Estimate propensity and prognostic scores.
}
\examples{
gen_mod <- generate_data(n = 100, 
                         dgp = 'ks', 
                         correct_outcome = FALSE,
                         correct_ps = TRUE)
this_data <- gen_mod$data
this_data <- estimate_scores(this_data, outcome_fm = outcome_fm,
                             ps_fm = ps_fm,
                             ps_fam = ps_fam,
                             outcome_fam = outcome_fam)
 this_data
}
