% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_fns.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for}
\usage{
generate_data(n, dgp, correct_outcome = TRUE, correct_ps = TRUE)
}
\arguments{
\item{n}{desired sample size.}

\item{dgp}{character string indicating desired data-generating process. Currently the following DGPs are implemented, each based on a different publication: \code{ks}, Kang and Shafer; \code{ld}, Lunceford and Davidian; \code{fi} Fan and Imai; \code{ls}, Leacy and Stuart; \code{ik} Iacus and King; \code{iw} Waernbaum; \code{pa} Austin.}

\item{correct_outcome}{logical indicating whether the outcome model should eb correctly specified.}

\item{correct_ps}{logical indicating whether the propensity score model should be correctly specified.}
}
\value{
a list of objects including
}
\description{
Generate data for
}
\examples{
gen_mod <- generate_data(n = 100, 
                         dgp = 'ks', 
                         correct_outcome = FALSE,
                         correct_ps = TRUE)
}
