% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_dl.R
\name{generate_data_dl}
\alias{generate_data_dl}
\title{Generate data for compliance using the Ding and Lu model.}
\usage{
generate_data_dl(n = 500, theta)
}
\arguments{
\item{n}{sample size}

\item{theta}{paramater controlling the violation of principal ignorability}
}
\value{
data.frame including \code{id}, covariate \code{x}, randomization indicator \code{z}, complier indicator \code{c}, treatment receipt indicator \code{s}, and outcome \code{y}.
}
\description{
Generate data for compliance using the Ding and Lu model.
}
\examples{
generate_data_dl(n = 500, theta = 1)
}
