% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_psr.R
\name{generate_data_psr}
\alias{generate_data_psr}
\title{Generate data for compliance under pseudo-randomization.}
\usage{
generate_data_psr(n, p, delta = 0.5, compliance_p = 0.75)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of covariates}

\item{delta}{complier average causal effect}

\item{compliance_p}{compliance proportion}
}
\value{
data.frame including \code{id}, treatment assignment indicator \code{z}, complier indicator \code{c}, treatment receipt indicator \code{s}, outcome \code{y}, and covariates \code{x1, ..., xp}.
}
\description{
Generate data for compliance under pseudo-randomization.
}
\examples{
generate_data_psr(n = 100, p = 8)
}
