% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_sj.R
\name{generate_data_sj}
\alias{generate_data_sj}
\title{Generate data for compliance using the Stuart and Jo model.}
\usage{
generate_data_sj(
  n,
  gamma_c = 0.5,
  compliance_p = 0.5,
  compliance_effect = 0.5,
  alpha_c = 1,
  alpha_n = 0.5,
  lambda_c = 1,
  lambda_n = 1,
  gamma_n = 0,
  sigma_c = 1,
  sigma_n = 1
)
}
\arguments{
\item{gamma_c}{effect of randomization among the compliers}

\item{compliance_p}{proportion of compliance}

\item{compliance_effect}{effect of covariate on compliance}

\item{alpha_c}{intercept term for compliers}

\item{alpha_n}{intercept term for never-takers}

\item{lambda_c}{covariate effect among compliers}

\item{lambda_n}{covariate effect among never-takers}

\item{gamma_n}{effect of randomization among never-takers}

\item{sigma_c}{error variance among compliers}

\item{sigma_n}{error variance among never-takers}
}
\value{
data.frame including \code{id}, covariate \code{x}, randomization indicator \code{z}, complier indicator \code{c}, treatment receipt indicator \code{s}, and outcome \code{y}.
}
\description{
Generate data for compliance using the Stuart and Jo model.
}
\examples{
generate_data_sj(n = 100)
}
