% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synthetic_estimator.R
\name{get_synthetic_estimator}
\alias{get_synthetic_estimator}
\title{Compute synthetic estimator. This is a wrapper function for combine_estimators.}
\usage{
get_synthetic_estimator(theta, boot, theta_0 = NULL, ...)
}
\arguments{
\item{theta}{vector or one-row data frame of candidate estimators.}

\item{boot}{matrix or data frame of bootstrapped versions of candidate estimators.}

\item{theta_0}{name of estimator presumed to be unbiased.}

\item{...}{additional arguments to be passed to \code{combine_estimators}.}
}
\value{
vector of synthetic estimators
}
\description{
Compute synthetic estimator. This is a wrapper function for combine_estimators.
}
