% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.R
\name{qp}
\alias{qp}
\title{Solve quadratic program}
\usage{
qp(qq, n_ests)
}
\arguments{
\item{qq}{objective function}

\item{n_ests}{number of candidate estimators}
}
\value{
list of objects from quadprog::solve.QP
}
\description{
Solve quadratic program
}
\examples{
rho <- 0.5
d <- 3
cov_mat <- rho*matrix(1, d, d) + (1-rho)*diag(d)
bias <- runif(d)
mse <- cov_mat + bias \%*\% t(bias)
qp(mse, d)
}
