% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_joint_distn.R
\name{resample_fn}
\alias{resample_fn}
\title{Resample ATEs.}
\usage{
resample_fn(dat, ypredfn = NULL, dpredfn, B, ate_list, outcome_fm, ps_fm,
  ps_fam, outcome_fam, cov_ids)
}
\arguments{
\item{dat}{data frame for analysis.}

\item{ypredfn}{function to predict y, mostly deprecated, default is NULL.}

\item{dpredfn}{function to predict the ATE used to demean data.}

\item{B}{number of bootstrap replications.}

\item{ate_list}{list of ATE functions.}

\item{outcome_fm}{outcome formula.}

\item{ps_fm}{propensity score formula.}

\item{ps_fam}{propensity score family.}

\item{outcome_fam}{outcome family.}

\item{cov_ids}{names of covariates for matching/balancing.}
}
\value{
list of tibbles of resampled data.
}
\description{
Resample ATEs.
}
